refactor: build every download's Content-Disposition through one helper
`contentDisposition` moves to `src/server/http.ts` and the subtraction
and analysis document downloads now go through it too, rather than
interpolating an unquoted `filename` parameter of their own.

Neither was reachable with a hostile name: a subtraction filename is
matched against a `subtraction_files` row the create job wrote, and the
analysis filename is a validated row id and a checked extension. Routing
them through the helper is so that a download added later against a
free-text name is not the one to discover the rule — which is what the
FASTA routes did.

The header is now quoted everywhere, which the subtraction test pins.
